requirements.txt
# Core Backend
fastapi
uvicorn
python-dotenv
requests

# AI & LangChain
google-generativeai
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers

# Data Processing
beautifulsoup4
lxml

Flask
python-dotenv
google-generativeai
requests